requirements.txt
jsbeautifier
matplotlib
nltk
scikit-image
tensorflow-gpu==1.13.1
tqdm
h5py==2.10.0
speaksee==0.0.1
Pillow==7.1.2